pull back

+++ src/main/java/com/niit/cmsdemo/service/impl/AdvertiseServiceImpl.java
@@ -5,6 +5,7 @@ import com.niit.cmsdemo.domain.Advertise;
 import com.niit.cmsdemo.service.AdvertiseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import java.util.List;
 import java.util.Map;
@@ -15,25 +16,26 @@ public class AdvertiseServiceImpl implements AdvertiseService {
     @Autowired
     private AdvertiseDao advertiseDao;
 
+
     @Override
-    public void addAdvertise(Advertise advertise) {
+    public void addAdvertise(Advertise advertise, String userId) {
+        advertise.setUserId(userId);
         advertiseDao.insertOne(advertise);
     }
 
     @Override
-    public void delAdvertise(Long[] adIds) {
+    public void delAdvertise(Long[] adIds, String userId) {
         for(Long id:adIds){
-            advertiseDao.deleteOne(id);
+            //检查userId与Advertise是否对应
+            if(advertiseDao.selectOne(id).getUserId().equals(userId)){
+                advertiseDao.deleteOne(id);
+            }
         }
     }
 
     @Override
-    public List<Advertise> findAll() {
-        return advertiseDao.selectAll();
-    }
-
-    @Override
-    public List<Advertise> findConditions(Map<String, Object> map) {
+    public List<Advertise> findConditions(Map<String, Object> map, String userId) {
+        map.put("userId",userId);
         return advertiseDao.selectConditions(map);
     }
 }
